// bank-lending-frontend/src/app/services/models/user-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserRequest {
  addressFact: string;
  addressRegister: string;
  dateOfBirth: string;
  email: string;
  firstName: string;
  lastname: string;
  passportSerialNumber: string;
  phoneNumber: string;
  surName?: string;
}